backend/spark-processor/requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
shapely==2.0.1
python-dotenv==1.0.0
